@@ -19,7 +19,7 @@ license-files = { paths = [
 authors = [
     { name = "William P. Krekelberg", email = "wpk@nist.gov" },
 ]
-requires-python = ">=3.8,<3.13"
+requires-python = ">=3.9,<3.13"
 classifiers = [
     "Development Status :: 2 - Pre-Alpha",
     "Intended Audience :: Science/Research",
@@ -27,7 +27,6 @@ classifiers = [
     "Operating System :: OS Independent",
     "Programming Language :: Python",
     "Programming Language :: Python :: 3 :: Only",
-    "Programming Language :: Python :: 3.8",
     "Programming Language :: Python :: 3.9",
     "Programming Language :: Python :: 3.10",
     "Programming Language :: Python :: 3.11",
@@ -82,7 +81,7 @@ optional-dependencies.nbval = [
     "nbval",
 ]
 optional-dependencies.resample = [
-    "cmomy>=0.16.0",
+    "cmomy>=0.17.0",
 ]
 optional-dependencies.test = [
     "pytest",
@@ -527,7 +526,6 @@ envs = [
     "test-extras",
 ]
 python = [
-    "3.8",
     "3.9",
     "3.10",
     "3.11",
