@@ -2,8 +2,6 @@
 
 import sys
 
-import pandas as pd
-
 if sys.version_info < (3, 11):
     from typing_extensions import Self
 else:
@@ -15,12 +13,12 @@ if sys.version_info < (3, 10):
 else:
     from typing import Concatenate, ParamSpec, TypeAlias
 
-if sys.version_info < (3, 9):
-    IndexAny = pd.Index
-else:
-    from typing import Any
+from typing import TYPE_CHECKING, Any
 
-    IndexAny: TypeAlias = "pd.Index[Any]"  # type: ignore[misc, unused-ignore]  # get pd.Index working for python 3.8
+if TYPE_CHECKING:
+    import pandas as pd
+
+IndexAny: TypeAlias = "pd.Index[Any]"  # type: ignore[misc, unused-ignore]  # get pd.Index working for python 3.8
 
 
 __all__ = [
